Update style.css
Put a space for comment in line 1.
Put a space for comment in line 5.
Put a space for comment in line 12.
Put a space for comment in line 54.
Put a space for comment in line 60.
Put a space for comment in line 66.
Put a space for comment in line 72.
Put a space for comment in line 84.
Remove a space for comment in line 96.
Beginning  coding style for contact form for files "contact.html".