Update style.css
Line 50 - 52 : Space on the left of the codes lines added.

Line 78 : line removed.

Line 100 - 112 : content of the class "books" and the class "books" removed.

Line 140 - 146 : @media for min-width: 901px : class "books" and in this class : "display: flex;" and "justified-content: space-around;" added.

Line 148 - 155 : @media for max-width: 900px : class "books" and in this class : "display: flex;" and "flex-direction: column;" and "align-items: center;" added.